requirements.txt
albucore==0.0.23
albumentations==1.4.2
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
babel==2.17.0
bce-python-sdk==0.9.35
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
easyocr==1.7.2
einops==0.2.0
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
Flask==3.1.0
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Session==0.8.0
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
gdown==4.4.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.7
imageio==2.37.0
imgaug==0.4.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgspec==0.19.0
mysql-connector-python==9.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.24.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle-gpu==2.6.2.post120
pandas==2.2.3
pillow==10.2.0
pillow_heif==0.22.0
prefetch-generator==1.0.1
protobuf==3.20.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
rarfile==4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.64
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shapely==2.1.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.4
sympy==1.13.1
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.107
ultralytics-thop==2.0.14
urllib3==2.4.0
vietocr==0.3.13
visualdl==2.5.3
Werkzeug==3.1.3
XlsxWriter==3.2.3
